// 25032026/sce-front-auditoria/sce-front-auditoria/src/app/model/resoluciones/tabResolucionBean.ts
export class TabResolucionBean{
  id: number;
  procedencia: number;
  fechaResolucion: Date;
  numeroExpediente: string;
  numeroResolucion: string;
  tipoResolucion: number;
  estadoResolucion: string;
  estadoDigitalizacion: string;
  observacionDigitalizacion: string;
  numeroPaginas: number;
  activo: number;
  audUsuarioCreacion: string;
  audFechaCreacion: Date;
  audUsuarioModificacion: string;
  audFechaModificacion: Date;
  asignado: number;
  audUsuarioAsignado: string;
  audFechaAsignado: Date;
}